BugFix: Damage on Environmental Hazards
Fixed a bug where the player could take damage from an environmental hazard, but if they stayed on the hazard when the invincibility frames wore off. They would not take damage.

Now it uses ComponentHit instead of BeginOverlap. Hit is launched every time the player moves towards an object. But it is only launched once when landing. To check if the player is still on top the hazard by the end of the invincibility frames, a box shaped sweep check is done beneath the player.