Tuesday 18.04.2023 Evening Bugfixes
Fixed
1: Grapplesmack dog when close to player caused dog to dog collision to launch.
Grapplesmack now seems to work properly as it did before the dog collision changes.

2: Slight fix on dogs going beneath the player when air grappled.
On air grapple, check if the player is close to the ground. If they are, then they will not do anything, but a delegate for ground grappling an enemy is done automatically when they land.