Sunday 23.04.2023
Created a testing module for debug draw testing mechanics and their collisions.

Found out complex collision gives odd responses to Linetraces. Where the impact.normal points towards the actor, regardless of the actual normal of the surface. Which makes it useless for finding out if an object is within another by using a capsule Sweep/Trace.

Created IK controls for the control rig, but have not yet added it to the animations.

PickupSpawner for health pickups. To hopefully help players to not run out of health while playing. These spawner will continually spawn health pickups the entire lifetime of the game.