Ledgegrab root anim and dog bounce shroom fix
Ledgegrab now drags the player up if their input is towards the ledge. The anim uses root motion to move the player character.

Dog bounce shroom fix for the watery cave tutorial dog. The dog shouldnt be able to fall off the shroom now and will spin properly throughout the bouncing.
The shrooms has a boolean that will forcefully set the bounced actors location to the arrow component. To make sure the bounce will 100% match where the shroom shows the bounce will go. Possibly making a finer version of it later.